create new component for current user infos
